fix: bundler outputs, tests coverage not working
