+++ src/types/DraggingConfiguration.ts
@@ -1,0 +1,13 @@
+export type DraggingConfiguration = Record<
+	| "chartDraggingDisabled"
+	| "datasetDraggingDisabled"
+	| "xAxisDraggingDisabled"
+	| "yAxisDraggingDisabled"
+	| "dataPointDraggingDisabled",
+	boolean
+>;
+
+export type AxisDraggingConfiguration = Pick<
+	DraggingConfiguration,
+	"xAxisDraggingDisabled" | "yAxisDraggingDisabled"
+>;
