+++ src/types/index.ts
@@ -2,3 +2,4 @@ export * from "./ChartJSTypes";
 export * from "./Configuration";
 export * from "./EventTypes";
 export * from "./DragDataState";
+export * from "./DraggingConfiguration";
